requirements.txt
streamlit
Pillow
numpy
osmnx
folium
geopy
pandas
streamlit-folium

